requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM Providers
openai==1.6.1
anthropic==0.8.1

# Vector Database
chromadb==0.4.18

# Embeddings
tiktoken==0.5.2

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
pypandoc==1.12

# Image Processing
Pillow==10.2.0
opencv-python==4.8.1.78

# Text Processing
nltk==3.8.1
spacy==3.7.2
beautifulsoup4==4.12.2

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Authentication & Security
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Web Framework (Optional - for future web interface)
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.29.0

# Utilities
requests==2.31.0
httpx==0.25.2
aiofiles==23.2.1
pyyaml==6.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1

